backend/requirements.txt
djangorestframework==3.14.0
psycopg2-binary==2.9.6
Pillow==9.5.0
djoser==2.2.0
urllib3==1.26.6
drf-extra-fields==3.5.0
django_filter==23.2
gunicorn==20.1.0
